Enhanced few of the features and added documentation

+++ src/main/java/com/retail/rewards/model/CreateTransactionRequest.java
@@ -3,14 +3,15 @@ package com.retail.rewards.model;
 import com.fasterxml.jackson.annotation.JsonFormat;
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 @Data
 public class CreateTransactionRequest {
 
-    private String phoneNumber;
-    private String name;
-    private Double billAmount;
-    @JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "yyyy-MM-dd")
-    private LocalDate transactionDate;
+	private String phoneNumber;
+	private String name;
+	private BigDecimal billAmount;
+	@JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "yyyy-MM-dd")
+	private LocalDate transactionDate;
 }
