+++ src/main/java/com/retail/rewards/model/RewardPointsResponse.java
@@ -2,9 +2,11 @@ package com.retail.rewards.model;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 public class RewardPointsResponse {
-    private String monthYear;
-    private long rewardPoints;
+	private String monthYear;
+	private BigDecimal rewardPoints;
 
 }
